requirements.txt
python-dotenv==1.0.1
pandas==2.2.2
Ipython==8.26.0
ipykernel==6.29.5
langchain==0.2.12
langgraph==0.2.3
langsmith==0.1.98
langchain-community==0.2.11
langchain-openai==0.1.20
langchain-chroma==0.1.2
tavily-python==0.3.7
